feat: Add project sync functionality to chatbotContext and syncData API
This commit adds the project sync functionality to the chatbotContext file and the syncData API route. In the chatbotContext file, a console.log statement is added to log the project before sending it to the database. In the syncData API route, a log statement is added to log the project before updating it in the database. This functionality allows for syncing project data between the frontend and backend. | added integration feature base line | added initial apis to do so